+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-04-16
+
+### Added
+- `strip_tags` helper for converting HTML to plain text
+- `:text_only` profile equivalent to `strip_tags`
+
 ## [0.2.3] - 2026-04-08
 
 ### Changed